+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file. This change
 ## [Unreleased]
 ### Added
 - Add a basic [Mandelbrot set] generator.
+- Add `draw-to-screen` for live viewing.
 
-[Unreleased]: https://github.com/your-name/mandelbrot/
+[Unreleased]: https://github.com/r-stuart/mandelbrot/
 [Mandelbrot set]: https://en.wikipedia.org/wiki/Mandelbrot_set